+++ tempest_fastapi_sdk/__init__.py
@@ -445,7 +445,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.168.2"
+__version__: str = "0.168.3"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",